+++ example/manual/gridfs/delete.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 import 'dart:io';
 
+import 'package:fixnum/fixnum.dart';
 import 'package:mongo_dart/mongo_dart.dart';
 
 const dbName = 'mongo-dart-example';
@@ -65,15 +66,15 @@ void main() async {
   // Set small chunks
   // This is normally don't needed. The default chunkSize is fine.
   // But we want to split are small data set into more than one chunks
-  GridFS.defaultChunkSize = 9;
+  GridFS.defaultChunkSize = Int32(9);
 
   // assures at least 3 chunks (for our tine test data set)
   var target = GridFS.defaultChunkSize * 3;
   var data = <int>[];
-  while (data.length < target) {
+  while (data.length < target.toInt()) {
     data.addAll(smallData);
   }
-  print('Expected chunks: ${(data.length / GridFS.defaultChunkSize).ceil()}');
+  print('Expected chunks: ${(data.length ~/ GridFS.defaultChunkSize.toInt())}');
   var extraData = <String, dynamic>{
     'test': [1, 2, 3],
     'extraData': 'Test',