+++ test/database_test.dart
@@ -1,5 +1,5 @@
 @Timeout(Duration(minutes: 10))
-library database_tests;
+library;
 
 import 'package:fixnum/fixnum.dart';
 import 'package:mongo_dart/mongo_dart.dart';