+++ test/message_test.dart
@@ -1,4 +1,4 @@
-library message_tests;
+library;
 
 import 'package:mongo_dart/mongo_dart.dart';
 import 'package:mongo_dart/src/database/message/mongo_modern_message.dart';