+++ test/mongo_dart_query_test.dart
@@ -1,4 +1,4 @@
-library mongo_dart_query_test;
+library;
 
 import 'package:test/test.dart';
 import 'package:bson/bson.dart';