+++ test/replica_tests.dart
@@ -1,4 +1,4 @@
-library replica_tests;
+library;
 
 import 'package:mongo_dart/mongo_dart.dart';
 import 'dart:async';