+++ test/test_binary_dump.dart
@@ -1,4 +1,4 @@
-library database_tests;
+library;
 
 import 'package:mongo_dart/mongo_dart.dart';
 import 'dart:io';